+++ app/commands/homebrew_context/daggerheart/change_race_command.rb
@@ -3,6 +3,10 @@
 module HomebrewContext
   module Daggerheart
     class ChangeRaceCommand < BaseCommand
+      include Deps[
+        refresh_user_data: 'services.homebrews_context.refresh_user_data'
+      ]
+
       use_contract do
         config.messages.namespace = :homebrew_race
 
@@ -16,6 +20,8 @@ module HomebrewContext
 
       def do_persist(input)
         input[:ancestry].update!(input.slice(:name))
+
+        refresh_user_data.call(user: input[:ancestry].user) if input[:ancestry].user
 
         { result: input[:ancestry] }
       end