+++ app/controllers/homebrews/feats_controller.rb
@@ -13,7 +13,7 @@ module Homebrews
     end
 
     def create
-      case add_feat.call(create_params.merge(user: current_user))
+      case add_feat.call(create_params.merge(user: current_user, bonuses: bonuses_params))
       in { errors: errors, errors_list: errors_list } then unprocessable_response(errors, errors_list)
       in { result: result } then serialize_resource(result, serializer, :feat, {}, :created)
       end