+++ app/serializers/homebrews/daggerheart/ancestry_serializer.rb
@@ -6,7 +6,16 @@ module Homebrews
       attributes :id, :name, :features
 
       def features
-        context && context[:features] ? context[:features].select { |i| i.origin_value == object.id } : []
+        return [] unless context
+        return [] unless context[:features]
+        return [] unless context[:bonuses]
+
+        resp = Panko::ArraySerializer.new(
+          context[:features].select { |item| item.origin_value == object.id },
+          each_serializer: Homebrews::FeatSerializer,
+          context: { bonuses: context[:bonuses] }
+        )
+        JSON.parse(resp.to_json)
       end
     end
   end