+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - separate web app for managing homebrews
+- bonuses for homebrews
 
 ## [0.3.15] - 2025-11-09
 ### Added